+++ packages/powersync-sdk-common/src/db/crud/SyncStatus.ts
@@ -26,7 +26,7 @@ export class SyncStatus {
    *  Currently this is reset to null after a restart.
    */
   get lastSyncedAt() {
-    return this.options.lastSyncedAt ?? null;
+    return this.options.lastSyncedAt;
   }
 
   /**
@@ -53,7 +53,7 @@ export class SyncStatus {
   }
 
   getMessage() {
-    const { dataFlow } = this.options;
+    const dataFlow = this.dataFlowStatus;
     return `SyncStatus<connected: ${this.connected} lastSyncedAt: ${this.lastSyncedAt}. Downloading: ${dataFlow.downloading}. Uploading: ${dataFlow.uploading}`;
   }