+++ demos/example-webpack/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [e472f17]
+  - @journeyapps/powersync-sdk-web@0.2.3
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [69592d0]
   - @journeyapps/powersync-sdk-web@0.2.2